+++ guillotina/const.py
@@ -1,0 +1,5 @@
+# base root object
+ROOT_ID = '0' * 32
+
+# object to reassign parent id to when you are deleting
+TRASHED_ID = 'D' * 32